Add files via upload
This version 14 has a better distributed merkle tree implementation and excludes the reputation management via quarks. It also excludes Quarks payments for transactions, it leaves the probabilistic one which ends up being the same. but eventually Quarks is better as it is more cash-flow efficient but also more complex for people to understand and implement.